align type-crate structure with the program's one
