+++ type-crate/src/types/interest_rate.rs
@@ -1,0 +1,104 @@
+use crate::assert_struct_size;
+
+#[cfg(feature = "anchor")]
+use {anchor_lang::prelude::*, type_layout::TypeLayout};
+
+use bytemuck::{Pod, Zeroable};
+
+#[cfg(not(feature = "anchor"))]
+use super::Pubkey;
+use super::WrappedI80F48;
+
+assert_struct_size!(InterestRateConfig, 240);
+#[repr(C)]
+#[cfg_attr(
+    feature = "anchor",
+    derive(AnchorDeserialize, AnchorSerialize, TypeLayout)
+)]
+#[derive(Debug, PartialEq, Eq, Copy, Clone, Pod, Zeroable, Default)]
+pub struct InterestRateConfig {
+    // Curve Params
+    pub optimal_utilization_rate: WrappedI80F48,
+    pub plateau_interest_rate: WrappedI80F48,
+    pub max_interest_rate: WrappedI80F48,
+
+    // Fees
+    /// Goes to insurance, funds `collected_insurance_fees_outstanding`
+    pub insurance_fee_fixed_apr: WrappedI80F48,
+    /// Goes to insurance, funds `collected_insurance_fees_outstanding`
+    pub insurance_ir_fee: WrappedI80F48,
+    /// Earned by the group, goes to `collected_group_fees_outstanding`
+    pub protocol_fixed_fee_apr: WrappedI80F48,
+    /// Earned by the group, goes to `collected_group_fees_outstanding`
+    pub protocol_ir_fee: WrappedI80F48,
+    pub protocol_origination_fee: WrappedI80F48,
+
+    pub _padding0: [u8; 16],
+    pub _padding1: [[u8; 32]; 3],
+}
+
+#[cfg_attr(
+    feature = "anchor",
+    derive(AnchorDeserialize, AnchorSerialize, TypeLayout)
+)]
+#[derive(Default, Clone, Debug, PartialEq, Eq)]
+pub struct InterestRateConfigOpt {
+    pub optimal_utilization_rate: Option<WrappedI80F48>,
+    pub plateau_interest_rate: Option<WrappedI80F48>,
+    pub max_interest_rate: Option<WrappedI80F48>,
+
+    pub insurance_fee_fixed_apr: Option<WrappedI80F48>,
+    pub insurance_ir_fee: Option<WrappedI80F48>,
+    pub protocol_fixed_fee_apr: Option<WrappedI80F48>,
+    pub protocol_ir_fee: Option<WrappedI80F48>,
+    pub protocol_origination_fee: Option<WrappedI80F48>,
+}
+
+#[repr(C)]
+#[cfg_attr(feature = "anchor", derive(AnchorDeserialize, AnchorSerialize))]
+#[derive(Default, Debug, PartialEq, Eq)]
+pub struct InterestRateConfigCompact {
+    // Curve Params
+    pub optimal_utilization_rate: WrappedI80F48,
+    pub plateau_interest_rate: WrappedI80F48,
+    pub max_interest_rate: WrappedI80F48,
+
+    // Fees
+    pub insurance_fee_fixed_apr: WrappedI80F48,
+    pub insurance_ir_fee: WrappedI80F48,
+    pub protocol_fixed_fee_apr: WrappedI80F48,
+    pub protocol_ir_fee: WrappedI80F48,
+    pub protocol_origination_fee: WrappedI80F48,
+}
+
+impl From<InterestRateConfigCompact> for InterestRateConfig {
+    fn from(ir_config: InterestRateConfigCompact) -> Self {
+        InterestRateConfig {
+            optimal_utilization_rate: ir_config.optimal_utilization_rate,
+            plateau_interest_rate: ir_config.plateau_interest_rate,
+            max_interest_rate: ir_config.max_interest_rate,
+            insurance_fee_fixed_apr: ir_config.insurance_fee_fixed_apr,
+            insurance_ir_fee: ir_config.insurance_ir_fee,
+            protocol_fixed_fee_apr: ir_config.protocol_fixed_fee_apr,
+            protocol_ir_fee: ir_config.protocol_ir_fee,
+            protocol_origination_fee: ir_config.protocol_origination_fee,
+            _padding0: [0; 16],
+            _padding1: [[0; 32]; 3],
+        }
+    }
+}
+
+impl From<InterestRateConfig> for InterestRateConfigCompact {
+    fn from(ir_config: InterestRateConfig) -> Self {
+        InterestRateConfigCompact {
+            optimal_utilization_rate: ir_config.optimal_utilization_rate,
+            plateau_interest_rate: ir_config.plateau_interest_rate,
+            max_interest_rate: ir_config.max_interest_rate,
+            insurance_fee_fixed_apr: ir_config.insurance_fee_fixed_apr,
+            insurance_ir_fee: ir_config.insurance_ir_fee,
+            protocol_fixed_fee_apr: ir_config.protocol_fixed_fee_apr,
+            protocol_ir_fee: ir_config.protocol_ir_fee,
+            protocol_origination_fee: ir_config.protocol_origination_fee,
+        }
+    }
+}
