+++ type-crate/src/types/mod.rs
@@ -1,9 +1,11 @@
 pub mod bank;
 pub mod bank_cache;
+pub mod bank_config;
 pub mod emode;
 pub mod fee_state;
 pub mod group;
 pub mod health_cache;
+pub mod interest_rate;
 pub mod pubkey;
 pub mod staked_settings;
 pub mod user_account;
@@ -11,10 +13,12 @@ pub mod wrapped_i80f48;
 
 pub use bank::*;
 pub use bank_cache::*;
+pub use bank_config::*;
 pub use emode::*;
 pub use fee_state::*;
 pub use group::*;
 pub use health_cache::*;
+pub use interest_rate::*;
 pub use pubkey::*;
 pub use staked_settings::*;
 pub use user_account::*;
